requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
python-dotenv
google-genai
